refactor: JS-version
