--- config/default.ts
@@ -1,5 +1,0 @@
-export default {
-  corsOrigin: "*",
-  port: 4000,
-  host: "localhost",
-};
