+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "server",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "src/app.js",
   "author": "",
   "license": "MIT",
   "scripts": {
-    "dev": "nodemon --config nodemon.json src/app.ts",
-    "start": "nodemon --config nodemon.json src/app.ts"
+    "dev": "nodemon --config nodemon.json src/app.js",
+    "start": "node src/app.js"
   },
   "devDependencies": {
     "@types/config": "^0.0.38",
@@ -25,5 +25,6 @@
     "pino": "^6.11.3",
     "pino-pretty": "^5.1.0",
     "socket.io": "^4.1.2"
-  }
+  },
+  "type": "module"
 }
